+++ apps/dashboard/app/(app)/ratelimits/[namespaceId]/logs/page.tsx
@@ -3,7 +3,6 @@ import { db } from "@/lib/db";
 import { notFound } from "next/navigation";
 
 import { CopyButton } from "@/components/dashboard/copy-button";
-import { EmptyPlaceholder } from "@/components/dashboard/empty-placeholder";
 import { Loading } from "@/components/dashboard/loading";
 import { Navbar as SubMenu } from "@/components/dashboard/navbar";
 import { Navbar } from "@/components/navbar";
@@ -20,7 +19,8 @@ import {
 } from "@/components/ui/table";
 import { clickhouse } from "@/lib/clickhouse";
 import { Gauge } from "@unkey/icons";
-import { Box, Check, X } from "lucide-react";
+import { Empty } from "@unkey/ui";
+import { Check, X } from "lucide-react";
 import { parseAsArrayOf, parseAsBoolean, parseAsIsoDateTime, parseAsString } from "nuqs/server";
 import { Suspense } from "react";
 import { navigation } from "../constants";
@@ -81,7 +81,7 @@ export default async function AuditPage(props: Props) {
             {namespace.name}
           </Navbar.Breadcrumbs.Link>
           <Navbar.Breadcrumbs.Link href={`/ratelimits/${props.params.namespaceId}/logs`} active>
-            Logs{" "}
+            Logs
           </Navbar.Breadcrumbs.Link>
         </Navbar.Breadcrumbs>
         <Navbar.Actions>
@@ -102,11 +102,9 @@ export default async function AuditPage(props: Props) {
           <Filters />
           <Suspense
             fallback={
-              <EmptyPlaceholder>
-                <EmptyPlaceholder.Icon>
-                  <Loading />
-                </EmptyPlaceholder.Icon>
-              </EmptyPlaceholder>
+              <div className="flex justify-center item-center mx-auto">
+                <Loading />
+              </div>
             }
           >
             <AuditLogTable
@@ -150,29 +148,27 @@ const AuditLogTable: React.FC<{
     country: selected.country.length > 0 ? selected.country : undefined,
     ipAddress: selected.ipAddress.length > 0 ? selected.ipAddress : undefined,
 
-    success: selected.success ?? undefined,
+    passed: selected.success ?? undefined,
   };
   const logs = await clickhouse.ratelimits.logs(query).then((res) => res.val!);
 
   if (logs.length === 0) {
     return (
-      <EmptyPlaceholder>
-        <EmptyPlaceholder.Icon>
-          <Box />
-        </EmptyPlaceholder.Icon>
-        <EmptyPlaceholder.Title>No logs found</EmptyPlaceholder.Title>
+      <Empty>
+        <Empty.Icon />
+        <Empty.Title>No logs found</Empty.Title>
         {isFiltered ? (
           <div className="flex flex-col items-center gap-2">
-            <EmptyPlaceholder.Description>
-              No events matched these filters, try changing them.{" "}
-            </EmptyPlaceholder.Description>
+            <Empty.Description>
+              No events matched these filters, try changing them.
+            </Empty.Description>
           </div>
         ) : (
-          <EmptyPlaceholder.Description>
+          <Empty.Description>
             Create, update or delete something and come back again.
-          </EmptyPlaceholder.Description>
+          </Empty.Description>
         )}
-      </EmptyPlaceholder>
+      </Empty>
     );
   }