+++ go/pkg/api/handler.go
@@ -1,0 +1,5 @@
+package api
+
+import "github.com/unkeyed/unkey/go/pkg/api/session"
+
+type Handler[TRequest any, TResponse any] func(s session.Session[TRequest, TResponse]) error