+++ apps/dashboard/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @unkey/web
 
+## 0.1.42
+
+### Patch Changes
+
+- @unkey/ratelimit@0.5.3
+
+## 0.1.41
+
+### Patch Changes
+
+- @unkey/ratelimit@0.5.2
+
 ## 0.1.40
 
 ### Patch Changes